+++ backend/api/v1/setting_service.go
@@ -901,7 +901,7 @@ func validateTableMetadata(engine v1pb.Engine, tableMetadata *v1pb.TableMetadata
 	if err := checkDatabaseMetadata(storepb.Engine(engine), tempStoreSchemaMetadata); err != nil {
 		return errors.Wrap(err, "failed to check database metadata")
 	}
-	if _, err := schema.GetDesignSchema(storepb.Engine(engine), tempStoreSchemaMetadata); err != nil {
+	if _, err := schema.GetDatabaseDefinition(storepb.Engine(engine), schema.GetDefinitionContext{}, tempStoreSchemaMetadata); err != nil {
 		return errors.Wrap(err, "failed to transform database metadata to schema string")
 	}
 	return nil
