--- backend/plugin/schema/pg/common.go
@@ -1,11 +1,0 @@
-package pg
-
-import "strings"
-
-func escapePostgreSQLString(s string) string {
-	return strings.ReplaceAll(s, "'", "''")
-}
-
-func unescapePostgreSQLString(s string) string {
-	return strings.ReplaceAll(s, "''", "'")
-}
