+++ backend/plugin/schema/schema.go
@@ -11,9 +11,7 @@ import (
 
 var (
 	mux                            sync.Mutex
-	getDesignSchemas               = make(map[storepb.Engine]getDesignSchema)
 	checkColumnTypes               = make(map[storepb.Engine]checkColumnType)
-	stringifyTables                = make(map[storepb.Engine]stringifyTable)
 	getDatabaseDefinitions         = make(map[storepb.Engine]getDatabaseDefinition)
 	getSchemaDefinitions           = make(map[storepb.Engine]getSchemaDefinition)
 	getTableDefinitions            = make(map[storepb.Engine]getTableDefinition)
@@ -24,9 +22,7 @@ var (
 	getSequenceDefinitions         = make(map[storepb.Engine]getSequenceDefinition)
 )
 
-type getDesignSchema func(*storepb.DatabaseSchemaMetadata) (string, error)
 type checkColumnType func(string) bool
-type stringifyTable func(*storepb.TableMetadata) (string, error)
 type getDatabaseDefinition func(GetDefinitionContext, *storepb.DatabaseSchemaMetadata) (string, error)
 type getSchemaDefinition func(*storepb.SchemaMetadata) (string, error)
 type getTableDefinition func(string, *storepb.TableMetadata, []*storepb.SequenceMetadata) (string, error)
@@ -177,23 +173,6 @@ func GetDatabaseDefinition(engine storepb.Engine, ctx GetDefinitionContext, meta
 	return f(ctx, metadata)
 }
 
-func RegisterGetDesignSchema(engine storepb.Engine, f getDesignSchema) {
-	mux.Lock()
-	defer mux.Unlock()
-	if _, dup := getDesignSchemas[engine]; dup {
-		panic(fmt.Sprintf("Register called twice %s", engine))
-	}
-	getDesignSchemas[engine] = f
-}
-
-func GetDesignSchema(engine storepb.Engine, to *storepb.DatabaseSchemaMetadata) (string, error) {
-	f, ok := getDesignSchemas[engine]
-	if !ok {
-		return "", errors.Errorf("engine %s is not supported", engine)
-	}
-	return f(to)
-}
-
 func RegisterCheckColumnType(engine storepb.Engine, f checkColumnType) {
 	mux.Lock()
 	defer mux.Unlock()
@@ -209,21 +188,4 @@ func CheckColumnType(engine storepb.Engine, tp string) bool {
 		return false
 	}
 	return f(tp)
-}
-
-func RegisterStringifyTable(engine storepb.Engine, f stringifyTable) {
-	mux.Lock()
-	defer mux.Unlock()
-	if _, dup := stringifyTables[engine]; dup {
-		panic(fmt.Sprintf("Register called twice %s", engine))
-	}
-	stringifyTables[engine] = f
-}
-
-func StringifyTable(engine storepb.Engine, table *storepb.TableMetadata) (string, error) {
-	f, ok := stringifyTables[engine]
-	if !ok {
-		return "", errors.Errorf("engine %s is not supported", engine)
-	}
-	return f(table)
 }
